Demo_RAG_Fast_API_and_memory_Version_2/requirements.txt
langchain==0.3.25 
langchain-community==0.3.25 
langchain-google-genai==2.1.5 
langchain-neo4j==0.4.0 
langchain-experimental==0.3.4 
python-dotenv==1.0.1 
openai==1.18.0
streamlit==1.32.2
fastapi
uvicorn
httpx
requests
pydantic

# Optional (for local dev and logging)
rich